fix: Summary, step8, do not create summary for STAR

+++ Scripts/08_Extract_Counts/Summarize_Jobs.sh
@@ -7,6 +7,11 @@ set -o nounset
 # export all variables from Pipeline_Setup.conf
 eval "$(../00_Setup_Pipeline/01_Pipeline_Setup.py --export)"
 
+if [ ${DEFAULT_ALIGNER} -eq 1 ]; then
+    echo "Summarizing for STAR aligner is not working at the moment. STAR output reports are different from TopHat output reports."
+    exit 0
+fi
+
 # output directory
 summary_dir="$(pwd)/Job_Summary"
 rm -rf "${summary_dir}" && mkdir -p "${summary_dir}"
